ref :: Float로 변경

+++ src/main/java/project/software/domain/user/controller/dto/request/UpdateLensPowerRequest.java
@@ -7,5 +7,5 @@ import lombok.NoArgsConstructor;
 @Getter
 public class UpdateLensPowerRequest {
 
-    private Long lensPower;
+    private Float lensPower;
 }
